Apply stackNavigation
